backend/requirements.txt
django
pillow
djangorestframework
djoser
django-cors-headers